Exclude confusing characters from barcodes (PP-1644) (#428)
## Description

Only generate barcodes with characters in this set:
`23456789ABCDEFGHJKMNPRSTUVWXYZabcdefghjkmnpqrstuvwxyz`

## Motivation and Context

We use both letters and numbers when generating VLC barcodes, but some
letters resemble numbers which can be confusing for users. We should not
use such letters when generating a barcode (though they should still be
allowed as a prefix): O/o, I/i, L/l.

## How Has This Been Tested?

- Running tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.