Fix privacy URL references  (#467)
## Description

- Update `session["privacy_url"]` to library.privacy_url
- Make sure `session["privacy_url"]` is cleaned up

## Motivation and Context

I noticed while working on
https://github.com/ThePalaceProject/virtual-library-card/pull/465 that
we are storing the libraries `terms_conditions_url` in the session as
`privacy_url`.

I'm not sure this is every actually used, so we might just be able to
pull these all out of the session, but I wanted to at least fix it so it
contains what it says it contains while I noticed it.

## How Has This Been Tested?

- Running unit tests

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.