Fix the broken docker build. (#566)
This update was triggered by the following error in the logs: 2.051
Package mime-support is not available, but is referred to by another
package. 2.051 This may mean that the package is missing, has been
obsoleted, or 2.051 is only available from another source
2.051 However the following packages replace it:
2.051   media-types mailcap

## Description

The build is broken due to changing package names. This update attempts
to fix it.
<!--- Describe your changes -->

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.